qt: add show/hide toggle to password controls
this adds a new button which toggles the visibility of text in the text
edit widget, enabling users to see the password they are typing, if they
wish, or to see the existing password they have set.

a concern here might be that people using someone else's account can thus
see such passwords, but they can access it anyway since its in plaintext
within the config file; sensitive passwords simply should not be stored
this way.

todo: currently just using the "eye"/"sight" character (U+1F441), but
perhaps we will want an icon?