config: swap name param with item pointer
looking up config items by name is inefficient, even with the bsearch
implementation; there are cases where we already (or could already) have
the actual pointer to hand.

this converts various config functions to take an item pointer, removing
the name lookup, placing the burden of performing name lookups on the
caller, though that can be hidden behind a macro.

note that
 1) while most functions hid the real config items from users by taking
    config names, the real items were not totally hidden from users since
    vlc_config_FindItem() (formerly config_FindConfig()) gives direct
    access to the pointer.
 2) there are cases where options are defined more than once and thus
    multiple copies exist, which actually means that reading/writing by name
    can involve the wrong instance, which can be avoided through item
    pointer use.