config: move module selection item cap attr storage
items of module-selection type hold the module capability they target in
string form, using the confusingly named 'psz_type' config attribute. this
is the only (current) use of this attribute. (did this attribute previously
have a different purpose?)

this is unnecessary; the min and max attributes are both capable of holding
string types, and both are unused for such item types, thus this attribute
is making all config items larger than they need to be, as well as
introducing confusion.

this moves storage of this value to min.psz, ditching the psz_type pointer

(of course the repurposing of min/max like this is also somewhat confusing,
but it is already done elsewhere - e.g. by subcat hint items)