config: unify subcat hint item storage
the subcat ID of a subcat hint config item is stored both in the value.i
attribute and the orig.i one, due to lack of clarity and inconsistency.
here we resolve this.

 - both Qt and MacOSX read it from value.i
 - plugin cache read/write uses orig.i

this switches the plugin cache over to using value.i, and removes the
duplicate storage in orig.i.