variables: do not reject empty string option values
a null psz_value indicates that the in-same-argument option name/value
separator ('=') was not found and thus in the case of data-taking (non-
bool) options this is invalid use of the option. it is thus right to
skip to cleanup in such a case.

however, i do not agree that the same should be done for all cases of an
empty value; how else do you override the default for a string type
option with an empty string value (which is a perfectly acceptable value
for a string type to have)?

this relaxes the check to allow empty string values through.