qt: improve key selector item identification
key selector config item names beginning with "key-" is a convention, not
something that is fixed/guaranteed. "global-" being prefix to the global
variants however is, since this is added automatically by the macro.

so, we should determine here the distinction between global/non-global by
presence, or not, of the "global-" prefix, to be robust.