qt: remove incorrect hotkey table control tooltip
the item pointer provided to the constructor is that of a single hotkey
config item (used in the advanced view this is that of the first hotkey
config item encountered, then all others are ignored; for simple view the
"key-play" option is arbitrary used).

using the longtext from this item in a tooltip on the label for the entire
hotkey table makes zero sense, so let's remove it.