qt: remove useless KeySelectorControl param
as just explained in removing the incorrect tooltip from the hotkey table's
label, the item pointer provided to the constructor of this control/widget
is that of a single hotkey config item (used in the advanced view this is
that of the first hotkey config item encountered, then all others are
ignored; for simple view the "key-play" option is arbitrary used). the
control contrastingly is used to edit the whole set of hotkey items.

the point in passing along an item pointer was that the base control class
requires one (all other controls target a single config item); it makes
absolutely no sense for this one though. this removes the param, and
passes along null to the base control constructor.