qt: restore focus for hotkey edit dialog buttons
This reverts commit d9b28c03a7b81dbf45ab74ec0fe6669de5aa173e.

The disabling of focus was done due to the design of the dialog
preventing use of the space button in key/sequence assignments
otherwise, as pressing the space bar triggered whichever button
was in focus.

This is now unnecessary with the redesign of this dialog based upon
QKeySequenceEdit, which can hold focus.