qt: clear hotkey conflict flag on fresh check
if the edit dialog detects a conflict, the 'conflicts' property is set to
true along with the dialog being updated to warn the user, offering them
the option of forcing a change in assignment. if they choose to force
assignment then that flag will correctly cause existing use to be cleared.

however, also available to the user is the option of simply entering a
different key/combination to be set, which will trigger a fresh conflict
check, and either this new combination will be accepted, if no conflict,
or the dialog will be updated with new conflict info. in the former case,
the 'conflict' property was left set, thus causing the table to be
pointlessly iterated over to clear existing use, when we know there isn't
any.

this change simply clears the 'conflict' attribute at the start of
performing a conflict check, thus preventing that.