qt: do not save config directly from first run handler
addresses fixme from 2011: d9dd2df25aa4a100e0bf580b44a5f2c19474d45d

during (clean) shutdown the config set, if dirty, is flushed to disk
(in libvlc_InternalCleanup()), so long as --ignore-config is not used.

thus, removing this direct save:

 1) does not stop the changes being saved as they currently are, so long
    as VLC does not crash.
 2) makes things more efficient if the user does also edit and save their
    preferences on first use, since these changes will be flushed in the
    same save action, rather than there being 2+ saves.
 3) removing this direct save means that --ignore-config is actually
    obeyed.

i do not follow the argument in the comment about other plugins changing
items of their own, though this comes from a time when writes had been
modelled in a module-scope way. the config set is primarily a description
of available options, but also holds state, with that state being a
representation of the saved settings file, plus any updates waiting to be
flushed to it. the object variables system is used to capture run time
state, with 'get' requests falling back to config item state when no
specific runtime state is set. changes to the config state, except from
preference interfaces, are very rare, but are made specifically with the
intention to be flushed (saved) to the save file at some point. thus i
don't see the issue with such other-plugin changes being flushed here at
the same time as the qt changes were.

however, as explained above, the best thing is to not perform the save
here, as it is inefficient to do so, and does not obey --ignore-config.
the changes are not important enough to force a write to survive a non-
clean shutdown.