mono: fix bad option default
-1 was not in the choice list, this meant that saving (advanced)
preferences changed the value (to 0), and thus it appeared as modified
in the new Qt expert prefs view.

this add the missing choice list entry corresponding to the default.
i've made a guess at the right label and explanation to use based upon
what seems to be happening with a -1 value in stereo_to_mono() and
from bits of git history relating to it.