upnp: fix bad option default
"auto" was not in the choice list, this meant that saving (advanced)
preferences changed the value (to 0), and thus it appeared as modified
in the new Qt expert prefs view.

the mistake here was simply that the default needed to be "Auto", not
"auto", since this module seems to be favouring capitalised choice list
string values.