visual: fix bad option default
"flat" was not in the choice list, this meant that saving (advanced)
preferences changed the value (to 0), and thus it appeared as modified
in the new Qt expert prefs view.

i have switched the default to "none", which matches the fallback used in
window_get_param(), which describes this as falling back to "using
rectangular window (i.e. no window)".