x264: fix option lists & defaults
this fixes a known (and deliberate) regression introduced in an earlier
commit

the problem with using the arrays in the x264 header is that they are null
terminated, which results in blank combobox entries in prefs GUIs. it is
similarly problematic for --help output. (such arrays are used for both
values and labels).

a previous commit of mine changed the modules/entry.c code to ignore a null
value/label pair at the end of arrays, thus fixing the problem in cases
like x264_nal_hrd_names. however, in other cases, like those addressed
here, a null value is actually wanted, but just needed a proper label. such
entries are currently not available per the previous commit (regression).
here we fix this by adding custom copies of the relevant arrays, with
proper labels given to these legitimate 'null' entries.

static assertions are used to pick up any future changes to the relevant
x264 header's arrays.